backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
Pillow==9.0.0
python-dotenv==1.0.0
gunicorn==20.1.0
django-colorfield==0.9.0
django-filter==23.2
psycopg2-binary==2.9.3
drf-extra-fields==3.7.0